+++ backend/src/index.ts
@@ -1,8 +1,42 @@
-import app from './app';
+import 'dotenv/config';
+import express from 'express';
+import morgan from 'morgan';
+import helmet from 'helmet';
+import cors from 'cors';
+import MessageResponse from './interfaces/MessageResponse';
 
-const port = process.env.PORT || 5000;
+// import {
+//   KeywordsRoutes,
+//   ModelRoutes,
+//   GenerationRoutes,
+// } from './routes/index.js';
+import { NotFound, ErrorHandler } from './middlewares';
+
+// import './db/db.js';
+
+const app = express();
+
+app.use(morgan('common'));
+app.use(helmet());
+app.use(
+  cors({
+    origin: process.env.CORS_ORIGIN || '*',
+  }),
+);
+app.use(express.json());
+
+console.log('[StereoAI]');
+
+app.get<{}, MessageResponse>('/', (req, res) => {
+  res.json({ message: 'StereoAI' });
+});
+
+// app.use('/api/keywords', KeywordsRoutes);
+// app.use('/api/models', ModelRoutes);
+// app.use('/api/generations', GenerationRoutes);
+app.use(NotFound, ErrorHandler);
+
+const port = process.env.PORT || 1337;
 app.listen(port, () => {
-  /* eslint-disable no-console */
   console.log(`Listening: http://localhost:${port}`);
-  /* eslint-enable no-console */
 });
