+++ backend/src/middlewares/index.ts
@@ -1,0 +1,4 @@
+import NotFound from './notFound';
+import ErrorHandler from './errorHandler';
+
+export { NotFound, ErrorHandler };
