--- src/middleware/errorHandler.js
@@ -1,10 +1,0 @@
-const ErrorHandler = (error, req, res, next) => {
-  const statusCode = res.statusCode === 200 ? 500 : res.statusCode;
-  res.status(statusCode);
-  res.json({
-    message: error.message,
-    stack: process.env.NODE_ENV === 'production' ? '🚫' : error.stack,
-  });
-};
-
-export default ErrorHandler;
