--- src/middleware/index.js
@@ -1,4 +1,0 @@
-import NotFound from './notFound.js';
-import ErrorHandler from './errorHandler.js';
-
-export { NotFound, ErrorHandler };
