--- src/middleware/notFound.js
@@ -1,7 +1,0 @@
-const NotFound = (req, res, next) => {
-  const error = new Error(`Not Found :: ${req.originalUrl}`);
-  res.status(404);
-  next(error);
-};
-
-export default NotFound;
